fix: move clients from orders to core and ecommerce_caller
