+++ commerce_coordinator/apps/core/clients.py
@@ -3,7 +3,6 @@ API clients for services that manage orders.
 """
 import logging
 
-import requests
 from django.conf import settings
 from edx_rest_api_client.client import OAuthAPIClient
 
@@ -33,31 +32,3 @@ class BaseEdxOAuthClient:
     @property
     def oauth2_client_secret(self):
         return settings.BACKEND_SERVICE_EDX_OAUTH2_SECRET
-
-
-class EcommerceApiClient(BaseEdxOAuthClient):
-    """
-    API client for calls to the edX Ecommerce service.
-    """
-    api_base_url = str(settings.ECOMMERCE_URL) + '/api/v2/'
-
-    def get_orders(self, query_params):
-        """
-        Call ecommerce API overview endpoint for data about an order.
-
-        Arguments:
-            username: restrict to orders by this username
-        Returns:
-            dict: Dictionary represention of JSON returned from API
-
-        See sample response in tests.py
-
-        """
-        try:
-            endpoint = self.api_base_url + 'orders/'
-            response = self.client.get(endpoint, params=query_params)
-            response.raise_for_status()
-            return response.json()
-        except requests.exceptions.HTTPError as exc:
-            logger.exception(exc)
-            raise
