+++ commerce_coordinator/apps/ecommerce_caller/admin.py
@@ -1,0 +1,4 @@
+""" Admin configuration for ecommerce_caller models. """
+# from django.contrib import admin
+
+# Register your models here.
