+++ commerce_coordinator/apps/ecommerce_caller/apps.py
@@ -1,0 +1,10 @@
+"""
+App configuration for the ecommerce_caller app.
+"""
+
+from django.apps import AppConfig
+
+
+class EcommerceCallerConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'ecommerce_caller'
