+++ commerce_coordinator/apps/ecommerce_caller/models.py
@@ -1,0 +1,7 @@
+"""
+Models for the ecommerce_caller app.
+"""
+
+# from django.db import models
+
+# Create your models here.
