+++ commerce_coordinator/apps/ecommerce_caller/tests/test_clients.py
@@ -1,0 +1,7 @@
+"""
+Tests for the ecommerce_caller.clients app.
+"""
+
+# from django.test import TestCase
+
+# Create your tests here.
