+++ commerce_coordinator/apps/ecommerce_caller/tests/test_views.py
@@ -1,0 +1,7 @@
+"""
+Tests for the ecommerce_caller views.
+"""
+
+# from django.test import TestCase
+
+# Create your tests here.
