+++ commerce_coordinator/apps/ecommerce_caller/views.py
@@ -1,0 +1,7 @@
+"""
+Views for the ecommerce_caller app which implement the API endpoints or callback handlers.
+"""
+
+# from django.shortcuts import render
+
+# Create your views here.
