+++ commerce_coordinator/apps/orders/tests/test_clients.py
@@ -6,7 +6,7 @@ import logging
 from django.test import TestCase
 from mock import patch
 
-from commerce_coordinator.apps.orders.clients import EcommerceApiClient
+from commerce_coordinator.apps.ecommerce_caller.clients import EcommerceApiClient
 from commerce_coordinator.apps.orders.tests import ECOMMERCE_REQUEST_EXPECTED_RESPONSE, ORDER_HISTORY_GET_PARAMETERS
 
 logger = logging.getLogger(__name__)
@@ -17,7 +17,7 @@ class OrdersClientTests(TestCase):
     Verify endpoint availability for order retrieval endpoint(s)
     """
 
-    @patch('commerce_coordinator.apps.orders.clients.EcommerceApiClient.get_orders')
+    @patch('commerce_coordinator.apps.ecommerce_caller.clients.EcommerceApiClient.get_orders')
     def test_ecommerce_api_client(self, mock_response):
         """We can call the EcommerceApiClient successfully."""
         mock_response.return_value = ECOMMERCE_REQUEST_EXPECTED_RESPONSE
