+++ docs/commerce_coordinator.apps.ecommerce_caller.migrations.rst
@@ -1,0 +1,10 @@
+commerce\_coordinator.apps.ecommerce_caller.migrations package
+==============================================================
+
+Module contents
+---------------
+
+.. automodule:: commerce_coordinator.apps.ecommerce_caller.migrations
+   :members:
+   :undoc-members:
+   :show-inheritance:
