+++ docs/commerce_coordinator.apps.ecommerce_caller.tests.rst
@@ -1,0 +1,29 @@
+commerce\_coordinator.apps.ecommerce_caller.tests package
+=========================================================
+
+Submodules
+----------
+
+commerce\_coordinator.apps.ecommerce_caller.tests.test\_clients module
+----------------------------------------------------------------------
+
+.. automodule:: commerce_coordinator.apps.ecommerce_caller.tests.test_clients
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+commerce\_coordinator.apps.ecommerce_caller.tests.test\_views module
+--------------------------------------------------------------------
+
+.. automodule:: commerce_coordinator.apps.ecommerce_caller.tests.test_views
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: commerce_coordinator.apps.ecommerce_caller.tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
