+++ docs/commerce_coordinator.apps.rst
@@ -10,6 +10,7 @@ Subpackages
    commerce_coordinator.apps.api
    commerce_coordinator.apps.core
    commerce_coordinator.apps.demo_lms
+   commerce_coordinator.apps.ecommerce_caller
    commerce_coordinator.apps.orders
    commerce_coordinator.apps.titan
